requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
install==1.3.5
python-dotenv==1.0.1
requests==2.32.3
setuptools==70.0.0
urllib3==2.2.1
